requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.1.0
cmake==3.26.3
contourpy==1.0.7
cycler==0.11.0
DeepImageSearch==2.5
dj-database-url==2.0.0
Django==4.2
django-admin-extra-buttons==1.5.6
django-heroku==0.3.1
faiss-cpu==1.7.4
filelock==3.12.0
fonttools==4.39.3
fsspec==2023.4.0
gunicorn==20.1.0
huggingface-hub==0.14.1
idna==3.4
importlib-resources==5.12.0
Jinja2==3.1.2
kiwisolver==1.4.4
lit==16.0.2
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
psycopg2==2.9.6
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.29.0
six==1.16.0
sqlparse==0.4.4
sympy==1.11.1
timm==0.6.13
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
triton==2.0.0
typing-extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
whitenoise==6.4.0
zipp==3.15.0
